requirements.txt
 pandas
 plotly_express
 streamlit
 numpy
 f_oneway